+++ packages/flutter_meedu/example/lib/main.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_meedu/router.dart' as router;
 import 'package:flutter_meedu/state.dart';
 import 'package:meedu/state.dart';
+import 'package:meedu_example/pages/example_page.dart';
 import 'package:meedu_example/routes/pages.dart';
 import 'package:meedu_example/routes/routes.dart';
 
@@ -26,12 +27,12 @@ class MyApp extends StatelessWidget {
           themeMode: controller.themeMode,
           theme: ThemeData.light(),
           darkTheme: ThemeData.dark(),
-          initialRoute: Routes.SPLASH,
-          // home: ProviderWidgetPage(),
+          // initialRoute: Routes.SPLASH,
+          home: ExamplePage(),
           navigatorObservers: [
             router.observer,
           ],
-          routes: Pages.routes,
+          // routes: Pages.routes,
         ),
       ),
       providers: [