+++ packages/flutter_meedu/lib/src/state/widgets/consumer_widget/consumer_widget.dart
@@ -9,8 +9,7 @@ part 'provider_filter.dart';
 /// [provider] must be a SimpleProvider or a BaseProvider.
 ///
 /// [filter] one instance of SimpleFilter or StateFilter, use this to avoid unnecessary  updates
-typedef ScopedReader = T Function<T, S>(BaseProvider<T> provider,
-    [WatchFilter<T, S>? filter]);
+typedef ScopedReader = T Function<T, S>(BaseProvider<T> provider, [WatchFilter<T, S>? filter]);
 
 class WatchFilter<T, S> {
   final BuildWhen<S>? when;
@@ -61,6 +60,7 @@ class _ConsumerState extends State<ConsumerWidget> {
   @override
   void initState() {
     super.initState();
+    print("consumer init");
     WidgetsBinding.instance!.addPostFrameCallback((_) {
       _afterFirstLayout = true;
     });